Changes hotkey mapping
The problem was that `space` and `enter` hotkeys are not the best
hotkeys to use as they overwrite hotkeys that are commonly used in
browser interactions for other things. The search bar can now be
focused with `alt` and random play can be triggered with `shift`.
Through this change it's now possible to hit space or enter to
(re)trigger a selected button.

Signed-off-by: Stephan Müller <smueller@suse.com>