Stop files individually
So a lot has changed to implement this feature. Now it's possible to
stop files on a file base level. This does also include search buttons,
which was a bit hard to implement, but hey it works great :)

Now each button has a stop button to it's right which is disabled and
gray if it can't be used. The button will be turned into active while
playing and the stop button will be colored red and be available to use.

Resolves #1
Signed-off-by: Stephan Müller <smueller@suse.com>