Fixes live-server start on windows
Added the direct path in order to be able to use the pack command to
have a full working environment, but this breaks the live-server usage
for windows users as their path structure is different by using "\"
instead of "/".

So if you use "npm pack" and extract that you have now to manually run
the following two commands "npm run updateSounds" and
"./node_modules/live-server/live-server.js" as "npm start" is not working as
"live-server" will not be found for some reason.

Signed-off-by: Stephan Müller <smueller@suse.com>